Ajustando o tipo das variáveis para ficar congruente com o resto do sistema

+++ GlobalGameJam2020/Assets/Scripts/UI/ScoreManager.cs
@@ -14,7 +14,7 @@ public class ScoreManager : MonoBehaviour
         scoreText.text = score.ToString();
     }
 
-    public void IncreaseScore(int amount, int hpAmount)
+    public void IncreaseScore(int amount, float hpAmount)
     {
         if(hpAmount <= 80)
         {
